logger bug fix

+++ nebixbot/log/logger.py
@@ -1,7 +1,6 @@
 import sys
 import logging
 import os
-import shutil
 
 from nebixbot import log
 
@@ -73,9 +72,8 @@ def delete_all_logs() -> bool:
         for filename in os.listdir(logsfile_dir):
             file_path = os.path.join(logsfile_dir, filename)
             if os.path.isfile(file_path) or os.path.islink(file_path):
-                os.unlink(file_path)
-            elif os.path.isdir(file_path):
-                shutil.rmtree(file_path)
+                if file_path.endswith('.log'):
+                    os.unlink(file_path)
         return True
     except Exception as err:
         print(f'Failed to delete {file_path}: {err}')
